+++ jipipe-r/src/main/java/org/hkijena/jipipe/extensions/r/algorithms/RScriptAlgorithm.java
@@ -61,6 +61,7 @@ public class RScriptAlgorithm extends JIPipeParameterSlotAlgorithm {
     private RScriptParameter script = new RScriptParameter();
     private RCaller rCaller;
     private JIPipeDynamicParameterCollection variables = new JIPipeDynamicParameterCollection(RUtils.ALLOWED_PARAMETER_CLASSES);
+    private boolean cleanUpAfterwards = true;
 
     public RScriptAlgorithm(JIPipeNodeInfo info) {
         super(info, JIPipeDefaultMutableSlotConfiguration.builder().build());
@@ -71,7 +72,20 @@ public class RScriptAlgorithm extends JIPipeParameterSlotAlgorithm {
         super(other);
         this.script = new RScriptParameter(other.script);
         this.variables = new JIPipeDynamicParameterCollection(other.variables);
+        this.cleanUpAfterwards = other.cleanUpAfterwards;
         registerSubParameter(variables);
+    }
+
+    @JIPipeDocumentation(name = "Clean up data after processing", description = "If enabled, data is deleted from temporary directories after " +
+            "the processing was finished. Disable this to make it possible to debug your scripts. The directories are accessible via the logs (Tools &gt; Logs).")
+    @JIPipeParameter("cleanup-afterwards")
+    public boolean isCleanUpAfterwards() {
+        return cleanUpAfterwards;
+    }
+
+    @JIPipeParameter("cleanup-afterwards")
+    public void setCleanUpAfterwards(boolean cleanUpAfterwards) {
+        this.cleanUpAfterwards = cleanUpAfterwards;
     }
 
     @Override
@@ -146,19 +160,21 @@ public class RScriptAlgorithm extends JIPipeParameterSlotAlgorithm {
         }
 
         // Clean up
-        progressInfo.log("Cleaning up ...");
-        for (Map.Entry<String, Path> entry : inputSlotPaths.entrySet()) {
-            try {
-                FileUtils.deleteDirectory(entry.getValue().toFile());
-            } catch (IOException e) {
-                e.printStackTrace();
+        if(cleanUpAfterwards) {
+            progressInfo.log("Cleaning up ...");
+            for (Map.Entry<String, Path> entry : inputSlotPaths.entrySet()) {
+                try {
+                    FileUtils.deleteDirectory(entry.getValue().toFile());
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
             }
-        }
-        for (Map.Entry<String, Path> entry : outputSlotPaths.entrySet()) {
-            try {
-                FileUtils.deleteDirectory(entry.getValue().toFile());
-            } catch (IOException e) {
-                e.printStackTrace();
+            for (Map.Entry<String, Path> entry : outputSlotPaths.entrySet()) {
+                try {
+                    FileUtils.deleteDirectory(entry.getValue().toFile());
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
             }
         }
     }
